feat: change project structure - create folders 'students' and 'subjects' and move components; create  modules shared and redux
