+++ src/app/components/students/student-page/student-page.component.scss
@@ -1,6 +1,7 @@
 span {
   position: relative;
   top: -4px;
+  font-weight: bold;
 }
 
 .buttons-container {
