+++ src/app/components/students/student-table/student-table.component.scss
