+++ src/app/components/subjects/subject-form/subject-form.component.html
@@ -3,7 +3,7 @@
   <app-form
     [label1]="'* Name'"
     [label2]="'* Teacher: Name Surname'"
-    [label3]="'Cabinet'"
+    [label3]="'Room'"
     [label4]="'Description'"
     [place1]="'Physics'"
     [place2]="'Ivan Mask'"
