+++ src/app/components/subjects/subject-marks-page/subject-marks-page.component.scss
@@ -1,7 +1,8 @@
-@import '../../../styles/button';
+@import '../../../../styles/button';
 
 h2 {
   margin: 20px;
+  margin-bottom: 0;
 }
 
 form {
@@ -25,8 +26,6 @@ form {
 }
 
 .btn-add {
-  position: relative;
-  box-sizing: border-box;
   min-width: 50px;
   height: 50px;
   margin: 10px 0;
