+++ src/app/components/subjects/subject-table/subject-table.component.scss
