+++ src/app/components/subjects/subjects-list-page/subjects-list-page.component.scss
